+++ jupyter-book/_toc.yml
@@ -43,6 +43,8 @@
     - file: overfit/cross_validation_index
       sections:
         - file: python_scripts/cross_validation_train_test
+        - file: python_scripts/cross_validation_validation_curve
+        - file: python_scripts/cross_validation_learning_curve
         - file: python_scripts/cross_validation_ex_01
         - file: python_scripts/cross_validation_sol_01
     - file: overfit/overfit_take_away
