+++ jupyter-book/interpretation/interpretation_quiz.md
@@ -5,8 +5,8 @@ With a same dataset, feature importance might differs if:
 
 - a) we use two different models
 - b) we use two different train/test split with a same model
-- c) we use a same model with a different set of hyper-parameters
-- d) we use a same model with the same set of hyper-parameters but a different
+- c) we use a same model with a different set of hyperparameters
+- d) we use a same model with the same set of hyperparameters but a different
   random_state
 ```
 
@@ -34,7 +34,7 @@ The feature importance provided by the scikit-learn random forest:
 - a) has bias for categorical feature
 - b) has bias for continuous (high cardinality) feature
 - c) is independent from the train/test split
-- d) is independent from the hyper-parameters
+- d) is independent from the hyperparameters
 ```
 
 ```{admonition} Question
