+++ python_scripts/cross_validation_train_test.py
@@ -169,9 +169,11 @@ cv_results = pd.DataFrame(cv_results)
 cv_results.head()
 
 # %% [markdown]
+# ```{tip}
 # By convention, scikit-learn model evaluation tools always use a convention
 # where "higher is better", this explains we used
 # `scoring="neg_mean_absolute_error"` (meaning "negative mean absolute error").
+# ```
 #
 # Let us revert the negation to get the actual error:
 
@@ -265,232 +267,6 @@ print(f"The standard deviation of the target is: {y.std():.2f} k$")
 # But in all cases, an error of 45 k\$ might be too large to automatically use
 # our model to tag house value without expert supervision.
 #
-# To better understand the performance of our model and maybe find insights on
-# how to improve it we will compare the generalization error with the empirical
-# error. Thus, we need to compute the error on the training set, which is
-# possible using the `cross_validate` function.
-
-# %%
-cv_results = cross_validate(regressor, X, y,
-                            cv=cv, scoring="neg_mean_absolute_error",
-                            return_train_score=True, n_jobs=2)
-cv_results = pd.DataFrame(cv_results)
-
-# %% [markdown]
-# We will select the train and test score and take the error instead.
-
-# %%
-scores = pd.DataFrame()
-scores[["train error", "test error"]] = -cv_results[
-    ["train_score", "test_score"]]
-sns.histplot(scores, bins=50)
-_ = plt.xlabel("Mean absolute error (k$)")
-
-# %% [markdown]
-# By plotting the distribution of the empirical and generalization errors, we
-# get information about whether our model is over-fitting, under-fitting (or
-# both at the same time).
-#
-# Here, we observe a **small empirical error** (actually zero), meaning that
-# the model is **not under-fitting**: it is flexible enough to capture any
-# variations present in the training set.
-#
-# However the **significantly larger generalization error** tells us that the
-# model is **over-fitting**: the model has memorized many variations of the
-# training set that could be considered "noisy" because they do not generalize
-# to help us make good prediction on the test set.
-#
-# Some model hyper-parameters are usually the key to go from a model that
-# underfits to a model that overfits, hopefully going through a region were we
-# can get a good balance between the two. We can acquire knowledge by plotting
-# a curve called the validation curve. This curve applies the above experiment
-# and varies the value of a hyper-parameter.
-#
-# For the decision tree, the `max_depth` the main parameter to control the
-# trade-off between under-fitting and over-fitting.
-
-# %%
-# %%time
-from sklearn.model_selection import validation_curve
-
-max_depth = [1, 5, 10, 15, 20, 25]
-train_scores, test_scores = validation_curve(
-    regressor, X, y, param_name="max_depth", param_range=max_depth,
-    cv=cv, scoring="neg_mean_absolute_error", n_jobs=2)
-train_errors, test_errors = -train_scores, -test_scores
-
-# %% [markdown]
-# Now that we collected the results, we will show the validation curve by
-# plotting the empirical and generalization errors (as well as their
-# deviations).
-
-# %%
-_, ax = plt.subplots()
-
-error_type = ["Empirical error", "Generalization error"]
-errors = [train_errors, test_errors]
-
-for name, err in zip(error_type, errors):
-    ax.plot(max_depth, err.mean(axis=1), linestyle="-.", label=name,
-            alpha=0.8)
-    ax.fill_between(max_depth, err.mean(axis=1) - err.std(axis=1),
-                    err.mean(axis=1) + err.std(axis=1), alpha=0.5,
-                    label=f"std. dev. {name.lower()}")
-
-ax.set_xticks(max_depth)
-ax.set_xlabel("Maximum depth of decision tree")
-ax.set_ylabel("Mean absolute error (k$)")
-ax.set_title("Validation curve for decision tree")
-_ = plt.legend(bbox_to_anchor=(1.05, 0.8), loc="upper left")
-
-# %% [markdown]
-# The validation curve can be divided into three areas:
-#
-# - For `max_depth < 10`, the decision tree underfits. The empirical error and
-#   therefore also the generalization error are both high. The model is too
-#   constrained and cannot capture much of the variability of the target
-#   variable.
-#
-# - The region around `max_depth = 10` corresponds to the parameter for which
-#   the decision tree generalizes the best. It is flexible enough to capture a
-#   fraction of the variability of the target that generalizes, while not
-#   memorizing all of the noise in the target.
-#
-# - For `max_depth > 10`, the decision tree overfits. The empirical error
-#   becomes very small, while the generalization error increases. In this
-#   region, the models captures too much of the noisy part of the variations of
-#   the target and this harms its ability to generalize well to test data.
-#
-# Note that for `max_depth = 10`, the model overfits a bit as there is a gap
-# between the empirical error and the generalization error. It can also
-# potentially underfit also a bit at the same time, because the empirical error
-# is still far from zero (more than 30 k\$), meaning that the model might
-# still be too constrained to model interesting parts of the data. However the
-# generalization error is minimal, and this is what really matters. This is the
-# best compromise we could reach by just tuning this parameter.
-#
-# We were lucky that the variance of the errors was small compared to their
-# respective values, and therefore the conclusions above are quite clear. This
-# is not necessarily always the case.
-#
-# We will now focus on one factor that can affect this variance, namely, the
-# size of the dataset:
-
-# %%
-y.size
-
-# %% [markdown]
-# Let's do an experiment and reduce the number of samples and repeat the
-# previous experiment. We will create a function that define a `ShuffleSplit`
-# and given a regressor and the data `X` and `y` will run a cross-validation.
-# The function will finally return the generalization error as a NumPy array.
-
-
-# %%
-def make_cv_analysis(regressor, X, y):
-    cv = ShuffleSplit(n_splits=10, test_size=0.2)
-    cv_results = cross_validate(regressor, X, y,
-                                cv=cv, scoring="neg_mean_absolute_error",
-                                return_train_score=True)
-    cv_results = pd.DataFrame(cv_results)
-    return (cv_results["test_score"] * -1).values
-
-
-# %% [markdown]
-# Now that we have a function to run each experiment, we will create an array
-# defining the number of samples for which we want to run the experiments.
-
-# %%
-sample_sizes = [100, 500, 1000, 5000, 10000, 15000, y.size]
-
-# %%
-import numpy as np
-
-# to make our results reproducible
-rng = np.random.RandomState(0)
-
-# create a dictionary where we will store the result of each run
-scores_sample_sizes = {"# samples": [], "test error": []}
-for n_samples in sample_sizes:
-    # select a subset of the data with a specific number of samples
-    sample_idx = rng.choice(np.arange(y.size), size=n_samples, replace=False)
-    X_sampled, y_sampled = X.iloc[sample_idx], y[sample_idx]
-    # run the experiment
-    score = make_cv_analysis(regressor, X_sampled, y_sampled)
-    # store the results
-    scores_sample_sizes["# samples"].append(n_samples)
-    scores_sample_sizes["test error"].append(score)
-
-# %% [markdown]
-# Now, we collected all our results and we will create a pandas dataframe to
-# easily make some plot.
-
-# %%
-scores_sample_sizes = pd.DataFrame(
-    np.array(scores_sample_sizes["test error"]).T,
-    columns=scores_sample_sizes["# samples"],
-)
-
-# %%
-sns.displot(scores_sample_sizes, kind="kde")
-plt.xlabel("Mean absolute error (k$)")
-_ = plt.title("Generalization errors distribution \n"
-              "by varying the sample size")
-
-# %% [markdown]
-# For the different sample sizes, we plotted the distribution of the
-# generalization error. We observe that smaller is the sample size; larger is
-# the variance of the generalization errors. Thus, having a small number of
-# samples might put us in a situation where it is impossible to get a reliable
-# evaluation.
-#
-# Here, we plotted the different curve to highlight the issue of small sample
-# size. However, this experiment is also used to draw the so-called **learning
-# curve**. This curve gives some additional indication regarding the benefit of
-# adding new training samples to improve a model's performance.
-
-# %%
-from sklearn.model_selection import learning_curve
-
-results = learning_curve(
-    regressor, X, y, train_sizes=sample_sizes[:-1], cv=cv,
-    scoring="neg_mean_absolute_error", n_jobs=2)
-train_size, train_scores, test_scores = results[:3]
-train_errors, test_errors = -train_scores, -test_scores
-
-# %% [markdown]
-# Now, we can plot the curve curve.
-
-# %%
-_, ax = plt.subplots()
-
-error_type = ["Empirical error", "Generalization error"]
-errors = [train_errors, test_errors]
-
-for name, err in zip(error_type, errors):
-    ax.plot(train_size, err.mean(axis=1), linestyle="-.", label=name,
-            alpha=0.8)
-    ax.fill_between(train_size, err.mean(axis=1) - err.std(axis=1),
-                    err.mean(axis=1) + err.std(axis=1),
-                    alpha=0.5, label=f"std. dev. {name.lower()}")
-
-ax.set_xticks(train_size)
-ax.set_xscale("log")
-ax.set_xlabel("Number of samples in the training set")
-ax.set_ylabel("Mean absolute error (k$)")
-ax.set_title("Learning curve for decision tree")
-_ = plt.legend(bbox_to_anchor=(1.05, 0.8), loc="upper left")
-
-# %% [markdown]
-# We see that the more samples we add to the training set on this learning
-# curve, the lower the error becomes. With this curve, we are searching for the
-# plateau for which there is no benefit to adding samples anymore or assessing
-# the potential gain of adding more samples into the training set.
-#
-# For this dataset we notice that our decision tree model would really benefit
-# from additional datapoints to reduce the amount of over-fitting and hopefully
-# reduce the generalization error even further.
-#
 # ## Summary
 #
 # In this notebook, we saw:
@@ -499,5 +275,3 @@ _ = plt.legend(bbox_to_anchor=(1.05, 0.8), loc="upper left")
 # * the meaning of the empirical and generalization errors;
 # * the overall cross-validation framework with the possibility to study
 #   performance variations;
-# * the effect of hyperparameter tuning using the validation curve;
-# * the effect of sample size using the learning curve.
